--- src/main/java/flow/RuntimeExceptionHandler.java
@@ -1,8 +1,0 @@
-package flow;
-
-/**
-* Created by aviadbendov on 1/1/15.
-*/
-public interface RuntimeExceptionHandler {
-    void handleException(RuntimeException ex);
-}
