+++ src/main/java/winter/flow/RuntimeExceptionHandler.java
@@ -1,0 +1,8 @@
+package winter.flow;
+
+/**
+* Created by aviadbendov on 1/1/15.
+*/
+public interface RuntimeExceptionHandler {
+    void handleException(RuntimeException ex);
+}
